refactor: 함수를 컴포넌트 내부로 이동
